bug fixed

+++ Irsa.PDM.Dtos/CertificacionFcMedios.cs
@@ -16,7 +16,7 @@ namespace Irsa.PDM.Dtos
         public DateTime fecha_aviso { get; set; }
         public string espacio { get; set; }
         public string des_tema { get; set; }
-        public int duracion_tema { get; set; }
+        public double duracion_tema { get; set; }
         public double costo_unitario { get; set; }
         public double descuento_1 { get; set; }
         public double descuento_2 { get; set; }
