+++ Irsa.PDM.Dtos/PautaItem.cs
@@ -7,7 +7,7 @@
         public string Espacio { get; set; }
         public string Tema { get; set; }      
         public double CostoUnitario { get; set; }
-        public int DuracionTema { get; set; }
+        public double DuracionTema { get; set; }
         public string Producto { get; set; }
         public double Descuento1 { get; set; }
         public double Descuento2 { get; set; }
