+++ Irsa.PDM.Entities/Certificacion.cs
@@ -13,7 +13,7 @@ namespace Irsa.PDM.Entities
         public DateTime? FechaAviso { get; set; }
         public string Espacio { get; set; }
         public string Tema { get; set; }
-        public int DuracionTema { get; set; }
+        public double DuracionTema { get; set; }
         public double CostoUnitario { get; set; }
         public double Descuento1 { get; set; }
         public double Descuento2 { get; set; }
